+++ src/util.hpp
@@ -1,0 +1,11 @@
+/* Copyright(c) Ryuichiro Nakato <rnakato@iam.u-tokyo.ac.jp>
+ * All rights reserved.
+ */
+#ifndef _UTIL_H_
+#define _UTIL_H_
+
+#include <string>
+
+std::string basename(const std::string &path);
+
+#endif /* _UTIL_H_ */